Fix #50: adapt pid file default path
Note: by default MDBF packages use a different path than official
distro's packages. IMO the behavior of distro is cleaner, so, make sure
that the directory exist in case we deploy MDBF packages.

Note2: the pid file option does not seem to be mandatory for systemd and
docker, so we might also decide to remove this option by default. But
for the moment it's probably cleaner/simpler to keep default options
deployed with packages.